requirements.txt
pandas
polygon-api-client
SQLAlchemy
python-dotenv
requests
psycopg2